Add new snooty driver docs to ingest (#410)
* add new snooty driver docs

* update source names

* add additional data sources + refactor and test snooty sources

* Apply suggestions from code review


* include landing now that fixed

---------
