+++ packages/ingest-mongodb-public/src/sources/snooty.test.ts
@@ -1,0 +1,14 @@
+import {
+  makeSnootyDataSources,
+  snootyProjectConfig,
+  snootyDataApiBaseUrl,
+} from "./snooty";
+
+describe("Snooty data sources", () => {
+  test.each(snootyProjectConfig)("$name should fetch data", async (project) => {
+    const [source] = await makeSnootyDataSources(snootyDataApiBaseUrl, [
+      project,
+    ]);
+    expect(source.name).toBeDefined();
+  });
+});
